fix(tests): import annotation utils from optional conftest utility
